ADD Authentication for seeing prices and purchasing from list of products

+++ store/models.py
@@ -5,7 +5,7 @@ from django.contrib.auth.models import User
 class Product(models.Model):
     """A model of a product being sold in a store"""
     name = models.CharField(max_length=50, default="")
-    price = models.FloatField(default=0.00)
+    price = models.PositiveIntegerField(default=0)
     image = models.CharField(max_length=200, default="")
 
     class Meta:
@@ -14,7 +14,7 @@ class Product(models.Model):
 
 class Catalog(models.Model):
     product = models.ForeignKey(Product, on_delete=models.PROTECT)
-    quantity = models.IntegerField(default=10)
+    quantity = models.PositiveIntegerField(default=10)
     selling_product = models.BooleanField(default=True)
 
 
@@ -27,4 +27,4 @@ class Order(models.Model):
 class OrderItem(models.Model):
     order = models.ForeignKey(Order, on_delete=models.PROTECT)
     product = models.OneToOneField(Product, on_delete=models.PROTECT)
-    quantity = models.IntegerField(default=1)
+    quantity = models.PositiveIntegerField(default=1)
